scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.2",
]
